requirements.txt
alembic==1.0.7
aniso8601==4.1.0
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.16.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
goodlogging==1.0.1
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.18.0
marshmallow-sqlalchemy==0.15.0
mccabe==0.6.1
pipenv==2018.11.14
psycopg2==2.7.7
psycopg2-binary==2.7.7
py-bcrypt==0.4
pygame==1.9.4
PyJWT==1.4.2
pylint==2.1.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
rarfile==2.7
regex==2018.11.22
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.17
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
Werkzeug==0.14.1
wrapt==1.10.11
